Fix calculating fat

+++ Dietician/Dietician/Models/ParameterResults.cs
@@ -7,7 +7,7 @@
         public double Cmp { get; set; }
         public double Protein { get; set; } 
         public double Carbohydrate { get; set; }
-        public int Fat { get; set; }
+        public double Fat { get; set; }
         public double Whr { get; set; }
         public string WhrLabel { get; set; }
         public double FatLevel { get; set; }
